refactor(job-manager): fold the github annotation sync into its only caller
syncGithubAnnotations is a method on ClusterClientSets with one call site, in the
resource event handler of cluster_client_sets.go, so it lived in a file of its own
for no reason a reader benefits from. Moving it next to that call site puts the
whole event-handling path in one place and leaves the file the third smallest in
the package.

Nothing else moves. syncer_status.go stays where it is: it is the persistence seam
three files write through, and it is the only place that decides whether a status
write is offloaded to the database, which should not be duplicated into a file
about pod events.
